Downloads PhenoCam time series

This is a wrapper around most of all the other functions. It downloads a time series and extract relevant phenological transition dates or phenophases.

download_phenocam(
  site = "harvard$",
  veg_type = NULL,
  frequency = "3",
  roi_id = NULL,
  outlier_detection = TRUE,
  smooth = TRUE,
  contract = FALSE,
  daymet = FALSE,
  trim_daymet = TRUE,
  trim = NULL,
  phenophase = FALSE,
  out_dir = tempdir(),
  internal = FALSE
)

Arguments

site
the site name, as mentioned on the PhenoCam web page expressed as a regular expression ("harvard$" == exact match)

veg_type
vegetation type (DB, EN, ... default = ALL)

frequency
frequency of the time series product (1, 3, "roistats")

roi_id
the id of the ROI to download (default = ALL)

outlier_detection
TRUE or FALSE, detect outliers

smooth
smooth data (logical, default is TRUE)

contract
contract 3-day data (logical, default is TRUE)

daymet
TRUE or FALSE, merges the daymet data

trim_daymet
TRUE or FALSE, trims data to match PhenoCam data

trim
year (numeric) to which to constrain the output (default = NULL)

phenophase
logical, calculate transition dates (default = FALSE)

out_dir
output directory where to store downloaded data (default = tempdir())

internal
allow for the data element to be returned to the workspace

Value

Downloaded files in out_dir of requested time series products, as well as derived phenophase estimates based upon these time series.

Examples

if (FALSE) {
# download the first ROI time series for the Harvard PhenoCam site
# at an aggregation frequency of 3-days.
download_phenocam(site = "harvard$",
                  veg_type = "DB",
                  roi_id = "1000",
                  frequency = "3")
 
# read phenocam data into phenocamr data structure                  
df <- read_phenocam(file.path(tempdir(),"harvard_DB_1000_3day.csv"))
                  
}