Calculates phenophases, for rising and falling parts of a time series

This routine combines a forward and backward run of transition_dates function to calculate the phenophases in both rising and falling parts of a PhenoCam time series.

phenophases(data, mat = NULL, internal = TRUE, out_dir = tempdir(),
  ...)

Arguments

data
a PhenoCam data file (or data frame)

mat
mean annual temperature

internal
return PhenoCam data file or data frame

out_dir
output directory

...
pass parameters to the transition_dates() function

Value

Estimates of transition dates for both rising and falling parts of a PhenoCam time series. All time series are evaluated (gcc_90, gcc_75, etc). The function returns a nested list with UNIX time based values including uncertainties on these estimates and their associated thresholds. When written to disk UNIX dates are converted to YYYY-MM-DD. The nested list has named locations rising and falling, or location 1 and 2 in the list respectivelly.

Examples

# downloads a time series
download_phenocam(site = "harvard$",
                  veg_type = "DB",
                  roi_id = "1000",
                  frequency = "3")

#> Downloading: harvard_DB_1000_3day.csv

#> -- Flagging outliers!

#> -- Smoothing time series!

# read in data as data frame and calculate phenophases
df <- read_phenocam(file.path(tempdir(),"harvard_DB_1000_3day.csv"))
my_dates <- phenophases(df, internal = TRUE)

# print results
print(my_dates)

#> $rising
#>    gcc_value transition_10 transition_25 transition_50 transition_10_lower_ci
#> 1     gcc_90         14002         14007         14014                  14000
#> 2     gcc_90         14361         14366         14374                  14358
#> 3     gcc_90         14719         14724         14730                  14715
#> 4     gcc_90         15096         15101         15107                  15093
#> 5     gcc_90         15452         15460         15467                  15444
#> 6     gcc_90         15825         15829         15835                  15823
#> 7     gcc_90         16196         16200         16207                  16195
#> 8     gcc_90         16556         16560         16566                  16554
#> 9     gcc_90         16927         16932         16939                  16925
#> 10    gcc_90         17287         17295         17302                  17283
#> 11    gcc_90         17653         17658         17664                  17651
#> 13    gcc_75         14002         14007         14014                  14001
#> 14    gcc_75         14360         14366         14374                  14357
#> 15    gcc_75         14720         14725         14731                  14716
#> 16    gcc_75         15095         15101         15107                  15091
#> 17    gcc_75         15451         15461         15468                  15443
#> 18    gcc_75         15824         15829         15835                  15823
#> 19    gcc_75         16196         16200         16207                  16195
#> 20    gcc_75         16556         16560         16566                  16554
#> 21    gcc_75         16928         16932         16939                  16926
#> 22    gcc_75         17287         17295         17302                  17283
#> 23    gcc_75         17654         17659         17665                  17652
#> 25    gcc_50         14003         14007         14014                  14001
#> 26    gcc_50         14359         14366         14374                  14352
#> 27    gcc_50         14720         14725         14731                  14703
#> 28    gcc_50         15095         15101         15109                  15092
#> 29    gcc_50         15448         15462         15469                  15442
#> 30    gcc_50         15824         15829         15836                  15821
#> 31    gcc_50         16196         16201         16207                  16195
#> 32    gcc_50         16556         16560         16566                  16553
#> 33    gcc_50         16927         16932         16939                  16925
#> 34    gcc_50         17286         17295         17302                  17282
#> 35    gcc_50         17654         17659         17665                  17652
#> 37  gcc_mean         14002         14007         14014                  14001
#> 38  gcc_mean         14358         14366         14374                  14351
#> 39  gcc_mean         14719         14725         14731                  14704
#> 40  gcc_mean         15095         15101         15108                  15092
#> 41  gcc_mean         15448         15462         15469                  15442
#> 42  gcc_mean         15824         15829         15835                  15821
#> 43  gcc_mean         16197         16201         16207                  16195
#> 44  gcc_mean         16556         16560         16566                  16554
#> 45  gcc_mean         16927         16932         16939                  16924
#> 46  gcc_mean         17286         17295         17302                  17282
#> 47  gcc_mean         17654         17658         17665                  17652
#>    transition_25_lower_ci transition_50_lower_ci transition_10_upper_ci
#> 1                   14006                  14012                  14004
#> 2                   14365                  14372                  14364
#> 3                   14723                  14728                  14722
#> 4                   15100                  15105                  15098
#> 5                   15459                  15465                  15456
#> 6                   15828                  15833                  15827
#> 7                   16199                  16205                  16198
#> 8                   16559                  16564                  16559
#> 9                   16931                  16937                  16930
#> 10                  17294                  17300                  17292
#> 11                  17657                  17662                  17656
#> 13                  14006                  14012                  14004
#> 14                  14365                  14372                  14363
#> 15                  14724                  14729                  14723
#> 16                  15100                  15106                  15098
#> 17                  15460                  15466                  15456
#> 18                  15828                  15834                  15827
#> 19                  16199                  16205                  16198
#> 20                  16559                  16564                  16558
#> 21                  16931                  16937                  16930
#> 22                  17294                  17300                  17292
#> 23                  17658                  17663                  17656
#> 25                  14006                  14012                  14005
#> 26                  14365                  14372                  14363
#> 27                  14724                  14730                  14723
#> 28                  15100                  15107                  15098
#> 29                  15461                  15467                  15456
#> 30                  15828                  15834                  15826
#> 31                  16200                  16206                  16198
#> 32                  16559                  16564                  16558
#> 33                  16931                  16938                  16930
#> 34                  17294                  17300                  17291
#> 35                  17658                  17664                  17657
#> 37                  14006                  14012                  14005
#> 38                  14365                  14372                  14362
#> 39                  14724                  14729                  14722
#> 40                  15100                  15107                  15097
#> 41                  15460                  15467                  15455
#> 42                  15828                  15834                  15826
#> 43                  16200                  16205                  16198
#> 44                  16559                  16564                  16558
#> 45                  16931                  16937                  16929
#> 46                  17294                  17300                  17290
#> 47                  17657                  17663                  17656
#>    transition_25_upper_ci transition_50_upper_ci threshold_10 threshold_25
#> 1                   14009                  14015      0.37778      0.39230
#> 2                   14369                  14375      0.37727      0.38876
#> 3                   14726                  14731      0.37749      0.39024
#> 4                   15103                  15108      0.37990      0.39440
#> 5                   15462                  15468      0.38294      0.39621
#> 6                   15831                  15836      0.38256      0.39734
#> 7                   16202                  16208      0.38173      0.39684
#> 8                   16562                  16567      0.38022      0.39308
#> 9                   16934                  16940      0.38436      0.39900
#> 10                  17297                  17303      0.38591      0.40013
#> 11                  17660                  17665      0.38393      0.39827
#> 13                  14009                  14015      0.37572      0.38945
#> 14                  14368                  14375      0.37441      0.38635
#> 15                  14727                  14732      0.37623      0.38974
#> 16                  15102                  15108      0.37736      0.39175
#> 17                  15463                  15469      0.38015      0.39378
#> 18                  15831                  15836      0.37844      0.39324
#> 19                  16202                  16208      0.37948      0.39311
#> 20                  16562                  16567      0.37850      0.39093
#> 21                  16934                  16940      0.38355      0.39577
#> 22                  17297                  17303      0.38277      0.39600
#> 23                  17660                  17666      0.38270      0.39786
#> 25                  14009                  14015      0.37480      0.38538
#> 26                  14369                  14375      0.37186      0.38322
#> 27                  14727                  14732      0.37360      0.38532
#> 28                  15103                  15110      0.37453      0.38667
#> 29                  15464                  15470      0.37721      0.39089
#> 30                  15831                  15837      0.37643      0.38894
#> 31                  16203                  16208      0.37628      0.39263
#> 32                  16561                  16567      0.37596      0.38796
#> 33                  16934                  16940      0.37852      0.39191
#> 34                  17297                  17303      0.37953      0.39189
#> 35                  17661                  17666      0.37938      0.39352
#> 37                  14009                  14015      0.37129      0.38491
#> 38                  14368                  14375      0.37050      0.38306
#> 39                  14727                  14732      0.37259      0.38595
#> 40                  15103                  15109      0.37378      0.38701
#> 41                  15463                  15470      0.37675      0.39141
#> 42                  15831                  15836      0.37574      0.38941
#> 43                  16202                  16208      0.37787      0.39259
#> 44                  16561                  16567      0.37522      0.38743
#> 45                  16934                  16940      0.37837      0.39216
#> 46                  17297                  17303      0.37895      0.39257
#> 47                  17660                  17666      0.37933      0.39017
#>    threshold_50 min_gcc max_gcc
#> 1       0.41959 0.36939 0.46404
#> 2       0.41847 0.36817 0.46226
#> 3       0.41821 0.36906 0.46466
#> 4       0.42098 0.37071 0.46827
#> 5       0.42737 0.37257 0.47722
#> 6       0.42526 0.37391 0.47088
#> 7       0.43235 0.37311 0.48027
#> 8       0.42025 0.37287 0.46648
#> 9       0.43000 0.37526 0.47893
#> 10      0.43733 0.37457 0.49268
#> 11      0.42962 0.37506 0.48006
#> 13      0.41534 0.36824 0.45799
#> 14      0.41440 0.36629 0.45654
#> 15      0.41732 0.36773 0.46123
#> 16      0.41536 0.36883 0.46182
#> 17      0.42234 0.37070 0.46851
#> 18      0.41914 0.37095 0.46506
#> 19      0.42661 0.37160 0.47374
#> 20      0.41726 0.37106 0.45989
#> 21      0.42614 0.37390 0.47420
#> 22      0.43099 0.37252 0.48171
#> 23      0.42595 0.37350 0.47280
#> 25      0.41109 0.36625 0.45179
#> 26      0.40938 0.36332 0.45276
#> 27      0.41078 0.36449 0.45612
#> 28      0.41629 0.36580 0.45846
#> 29      0.41893 0.36802 0.46492
#> 30      0.41703 0.36773 0.46090
#> 31      0.42059 0.36881 0.46879
#> 32      0.41363 0.36734 0.45505
#> 33      0.42140 0.37008 0.46996
#> 34      0.42445 0.36958 0.47377
#> 35      0.42051 0.37074 0.46939
#> 37      0.41148 0.36482 0.45093
#> 38      0.40999 0.36204 0.45164
#> 39      0.41188 0.36355 0.45559
#> 40      0.41195 0.36502 0.45677
#> 41      0.42018 0.36708 0.46473
#> 42      0.41400 0.36689 0.45974
#> 43      0.42136 0.36791 0.46820
#> 44      0.41319 0.36676 0.45478
#> 45      0.42199 0.36884 0.46934
#> 46      0.42604 0.36847 0.47609
#> 47      0.42180 0.36962 0.46782
#> 
#> $falling
#>    gcc_value transition_10 transition_25 transition_50 transition_10_lower_ci
#> 1     gcc_90         17839         17835         17825                  17841
#> 2     gcc_90         17473         17470         17460                  17474
#> 3     gcc_90         17102         17098         17091                  17103
#> 4     gcc_90         16738         16734         16726                  16739
#> 5     gcc_90         16364         16361         16355                  16365
#> 6     gcc_90         15995         15992         15981                  15997
#> 7     gcc_90         15632         15628         15621                  15634
#> 8     gcc_90         15275         15270         15258                  15276
#> 9     gcc_90         14900         14896         14889                  14901
#> 10    gcc_90         14540         14537         14530                  14541
#> 11    gcc_90         14176         14173         14166                  14177
#> 12    gcc_75         17839         17835         17824                  17840
#> 13    gcc_75         17473         17470         17461                  17474
#> 14    gcc_75         17102         17098         17090                  17103
#> 15    gcc_75         16735         16732         16725                  16736
#> 16    gcc_75         16364         16361         16354                  16365
#> 17    gcc_75         15995         15992         15983                  15997
#> 18    gcc_75         15630         15626         15618                  15631
#> 19    gcc_75         15275         15269         15257                  15276
#> 20    gcc_75         14899         14895         14888                  14900
#> 21    gcc_75         14540         14537         14530                  14541
#> 22    gcc_75         14175         14172         14165                  14176
#> 23    gcc_50         17837         17834         17826                  17838
#> 24    gcc_50         17473         17469         17461                  17474
#> 25    gcc_50         17101         17097         17088                  17102
#> 26    gcc_50         16733         16730         16724                  16734
#> 27    gcc_50         16363         16359         16352                  16364
#> 28    gcc_50         15995         15992         15983                  15997
#> 29    gcc_50         15629         15625         15616                  15630
#> 30    gcc_50         15274         15269         15257                  15276
#> 31    gcc_50         14900         14896         14887                  14901
#> 32    gcc_50         14539         14536         14528                  14540
#> 33    gcc_50         14175         14172         14165                  14176
#> 34  gcc_mean         17837         17834         17824                  17839
#> 35  gcc_mean         17473         17469         17461                  17474
#> 36  gcc_mean         17101         17097         17089                  17102
#> 37  gcc_mean         16733         16730         16724                  16734
#> 38  gcc_mean         16363         16359         16353                  16364
#> 39  gcc_mean         15995         15992         15983                  15997
#> 40  gcc_mean         15629         15625         15616                  15630
#> 41  gcc_mean         15275         15269         15258                  15276
#> 42  gcc_mean         14899         14895         14887                  14900
#> 43  gcc_mean         14539         14536         14528                  14540
#> 44  gcc_mean         14175         14172         14166                  14176
#>    transition_25_lower_ci transition_50_lower_ci transition_10_upper_ci
#> 1                   17837                  17829                  17837
#> 2                   17471                  17463                  17471
#> 3                   17099                  17093                  17100
#> 4                   16735                  16729                  16736
#> 5                   16362                  16357                  16362
#> 6                   15993                  15986                  15993
#> 7                   15629                  15623                  15630
#> 8                   15272                  15261                  15272
#> 9                   14897                  14891                  14897
#> 10                  14538                  14532                  14538
#> 11                  14174                  14168                  14174
#> 12                  17836                  17828                  17836
#> 13                  17471                  17463                  17471
#> 14                  17099                  17092                  17100
#> 15                  16733                  16727                  16733
#> 16                  16362                  16356                  16362
#> 17                  15993                  15986                  15993
#> 18                  15627                  15620                  15628
#> 19                  15271                  15260                  15272
#> 20                  14896                  14890                  14897
#> 21                  14538                  14532                  14539
#> 22                  14173                  14167                  14173
#> 23                  17835                  17829                  17835
#> 24                  17470                  17463                  17471
#> 25                  17098                  17091                  17099
#> 26                  16731                  16726                  16731
#> 27                  16360                  16354                  16361
#> 28                  15993                  15986                  15993
#> 29                  15626                  15619                  15627
#> 30                  15271                  15260                  15272
#> 31                  14897                  14889                  14898
#> 32                  14537                  14530                  14537
#> 33                  14173                  14167                  14173
#> 34                  17835                  17827                  17835
#> 35                  17470                  17463                  17471
#> 36                  17098                  17091                  17099
#> 37                  16731                  16726                  16732
#> 38                  16360                  16355                  16361
#> 39                  15993                  15986                  15993
#> 40                  15626                  15619                  15627
#> 41                  15271                  15261                  15272
#> 42                  14896                  14889                  14897
#> 43                  14537                  14530                  14537
#> 44                  14173                  14168                  14173
#>    transition_25_upper_ci transition_50_upper_ci threshold_10 threshold_25
#> 1                   17833                  17823      0.39521      0.40644
#> 2                   17467                  17458      0.38995      0.40164
#> 3                   17096                  17090      0.38562      0.39951
#> 4                   16732                  16725      0.38156      0.39618
#> 5                   16359                  16354      0.38488      0.39631
#> 6                   15990                  15974      0.38903      0.39936
#> 7                   15626                  15619      0.38093      0.39416
#> 8                   15267                  15255      0.38694      0.39965
#> 9                   14894                  14888      0.37730      0.39007
#> 10                  14535                  14528      0.37198      0.38478
#> 11                  14171                  14165      0.36886      0.38092
#> 12                  17832                  17822      0.39065      0.40163
#> 13                  17468                  17459      0.38586      0.39707
#> 14                  17096                  17089      0.38319      0.39643
#> 15                  16730                  16724      0.38036      0.39353
#> 16                  16359                  16353      0.38257      0.39364
#> 17                  15990                  15981      0.38483      0.39494
#> 18                  15624                  15616      0.38296      0.39612
#> 19                  15267                  15255      0.38368      0.39673
#> 20                  14893                  14886      0.37740      0.38931
#> 21                  14535                  14529      0.37023      0.38246
#> 22                  14170                  14164      0.36968      0.38175
#> 23                  17832                  17824      0.38245      0.39436
#> 24                  17467                  17460      0.38250      0.39646
#> 25                  17095                  17087      0.38133      0.39433
#> 26                  16728                  16723      0.38030      0.39188
#> 27                  16357                  16351      0.38150      0.39461
#> 28                  15990                  15981      0.38218      0.39179
#> 29                  15623                  15614      0.37983      0.39196
#> 30                  15267                  15255      0.38178      0.39196
#> 31                  14893                  14885      0.37168      0.38313
#> 32                  14534                  14527      0.37078      0.38190
#> 33                  14170                  14164      0.36558      0.37741
#> 34                  17832                  17821      0.38797      0.39715
#> 35                  17467                  17459      0.38219      0.39596
#> 36                  17095                  17088      0.38089      0.39429
#> 37                  16729                  16723      0.38188      0.39337
#> 38                  16357                  16352      0.38098      0.39419
#> 39                  15990                  15980      0.38191      0.39149
#> 40                  15623                  15613      0.38043      0.39267
#> 41                  15267                  15254      0.37944      0.39218
#> 42                  14893                  14885      0.37363      0.38528
#> 43                  14534                  14526      0.36989      0.38101
#> 44                  14170                  14165      0.36478      0.37683
#>    threshold_50 min_gcc max_gcc
#> 1       0.42666 0.38805 0.46300
#> 2       0.42726 0.38269 0.47056
#> 3       0.42236 0.37816 0.46497
#> 4       0.42046 0.37564 0.46133
#> 5       0.42288 0.37781 0.46585
#> 6       0.42314 0.38120 0.46338
#> 7       0.41692 0.37258 0.45975
#> 8       0.42113 0.38105 0.46080
#> 9       0.41407 0.37164 0.45296
#> 10      0.41154 0.36366 0.45733
#> 11      0.40806 0.36043 0.44983
#> 12      0.42337 0.38592 0.45852
#> 13      0.42153 0.37900 0.46279
#> 14      0.41878 0.37598 0.46041
#> 15      0.41834 0.37593 0.45514
#> 16      0.42062 0.37554 0.46041
#> 17      0.41774 0.37735 0.45580
#> 18      0.41649 0.37762 0.45399
#> 19      0.41687 0.37829 0.45319
#> 20      0.40968 0.37067 0.44792
#> 21      0.40740 0.36222 0.45109
#> 22      0.40514 0.36338 0.44441
#> 23      0.41624 0.37604 0.45530
#> 24      0.41804 0.37601 0.45791
#> 25      0.41792 0.37502 0.45631
#> 26      0.41561 0.37594 0.45159
#> 27      0.41818 0.37642 0.45561
#> 28      0.41382 0.37474 0.45129
#> 29      0.41298 0.37532 0.44937
#> 30      0.41214 0.37481 0.44765
#> 31      0.40591 0.36619 0.44420
#> 32      0.40676 0.36457 0.44569
#> 33      0.40146 0.35946 0.43957
#> 34      0.41758 0.38114 0.45380
#> 35      0.41704 0.37575 0.45677
#> 36      0.41586 0.37439 0.45564
#> 37      0.41529 0.37514 0.45030
#> 38      0.41587 0.37595 0.45487
#> 39      0.41308 0.37446 0.45129
#> 40      0.41221 0.37576 0.44810
#> 41      0.41036 0.37426 0.44632
#> 42      0.40640 0.36704 0.44332
#> 43      0.40521 0.36370 0.44440
#> 44      0.39860 0.35868 0.43720
#>