Calculates phenophases, for rising and falling parts of a time series

This routine combines a forward and backward run of transition_dates function to calculate the phenophases in both rising and falling parts of a PhenoCam time series.

phenophases(data, mat = NULL, internal = TRUE, out_dir = tempdir(),
  ...)

Arguments

data
a PhenoCam data file (or data frame)

mat
mean annual temperature

internal
return PhenoCam data file or data frame

out_dir
output directory

...
pass parameters to the transition_dates() function

Value

Estimates of transition dates for both rising and falling parts of a PhenoCam time series. All time series are evaluated (gcc_90, gcc_75, etc). The function returns a nested list with UNIX time based values including uncertainties on these estimates and their associated thresholds. When written to disk UNIX dates are converted to YYYY-MM-DD. The nested list has named locations rising and falling, or location 1 and 2 in the list respectivelly.

Examples

# downloads a time series
download_phenocam(site = "harvard$",
                  veg_type = "DB",
                  roi_id = "1000",
                  frequency = "3")

#> Downloading: harvard_DB_1000_3day.csv

#> -- Flagging outliers!

#> -- Smoothing time series!

# read in data as data frame and calculate phenophases
df <- read_phenocam(file.path(tempdir(),"harvard_DB_1000_3day.csv"))
my_dates <- phenophases(df, internal = TRUE)

# print results
print(my_dates)

#> $rising
#>    gcc_value transition_10 transition_25 transition_50 transition_10_lower_ci
#> 1     gcc_90         14003         14007         14013                  14001
#> 2     gcc_90         14361         14367         14374                  14357
#> 3     gcc_90         14720         14724         14730                  14717
#> 4     gcc_90         15096         15101         15107                  15091
#> 5     gcc_90         15451         15461         15467                  15444
#> 6     gcc_90         15826         15829         15835                  15824
#> 7     gcc_90         16197         16201         16206                  16195
#> 8     gcc_90         16556         16561         16566                  16554
#> 9     gcc_90         16927         16933         16939                  16925
#> 10    gcc_90         17287         17296         17302                  17283
#> 11    gcc_90         17654         17658         17665                  17652
#> 12    gcc_75         14003         14007         14014                  14001
#> 13    gcc_75         14360         14367         14374                  14356
#> 14    gcc_75         14721         14725         14731                  14718
#> 15    gcc_75         15095         15101         15107                  15090
#> 16    gcc_75         15451         15461         15468                  15443
#> 17    gcc_75         15825         15829         15835                  15823
#> 18    gcc_75         16197         16201         16206                  16196
#> 19    gcc_75         16556         16561         16566                  16554
#> 20    gcc_75         16928         16933         16939                  16926
#> 21    gcc_75         17287         17296         17302                  17283
#> 22    gcc_75         17654         17659         17665                  17652
#> 23    gcc_50         14003         14007         14014                  14002
#> 24    gcc_50         14358         14367         14374                  14352
#> 25    gcc_50         14721         14725         14731                  14702
#> 26    gcc_50         15095         15102         15109                  15092
#> 27    gcc_50         15448         15462         15469                  15442
#> 28    gcc_50         15824         15830         15836                  15821
#> 29    gcc_50         16197         16202         16207                  16196
#> 30    gcc_50         16555         16560         16566                  16553
#> 31    gcc_50         16928         16933         16939                  16926
#> 32    gcc_50         17286         17296         17302                  17282
#> 33    gcc_50         17654         17659         17666                  17652
#> 34  gcc_mean         14003         14007         14013                  14002
#> 35  gcc_mean         14357         14366         14374                  14350
#> 36  gcc_mean         14720         14725         14731                  14702
#> 37  gcc_mean         15095         15101         15109                  15092
#> 38  gcc_mean         15447         15462         15469                  15442
#> 39  gcc_mean         15824         15829         15835                  15821
#> 40  gcc_mean         16197         16201         16207                  16196
#> 41  gcc_mean         16555         16560         16566                  16553
#> 42  gcc_mean         16927         16933         16939                  16925
#> 43  gcc_mean         17286         17296         17302                  17282
#> 44  gcc_mean         17654         17659         17665                  17652
#>    transition_25_lower_ci transition_50_lower_ci transition_10_upper_ci
#> 1                   14006                  14011                  14005
#> 2                   14365                  14372                  14364
#> 3                   14723                  14729                  14723
#> 4                   15100                  15105                  15099
#> 5                   15460                  15466                  15457
#> 6                   15828                  15833                  15828
#> 7                   16200                  16205                  16199
#> 8                   16560                  16565                  16559
#> 9                   16932                  16937                  16930
#> 10                  17295                  17301                  17291
#> 11                  17657                  17663                  17656
#> 12                  14006                  14012                  14005
#> 13                  14366                  14372                  14364
#> 14                  14724                  14729                  14723
#> 15                  15100                  15106                  15098
#> 16                  15460                  15466                  15456
#> 17                  15828                  15834                  15827
#> 18                  16200                  16205                  16199
#> 19                  16560                  16564                  16558
#> 20                  16932                  16937                  16930
#> 21                  17295                  17301                  17291
#> 22                  17658                  17664                  17656
#> 23                  14006                  14012                  14005
#> 24                  14366                  14372                  14363
#> 25                  14724                  14730                  14723
#> 26                  15101                  15107                  15098
#> 27                  15461                  15467                  15456
#> 28                  15829                  15834                  15827
#> 29                  16201                  16205                  16199
#> 30                  16559                  16564                  16558
#> 31                  16932                  16937                  16930
#> 32                  17295                  17301                  17290
#> 33                  17658                  17664                  17656
#> 34                  14006                  14011                  14005
#> 35                  14365                  14372                  14362
#> 36                  14724                  14729                  14723
#> 37                  15100                  15107                  15098
#> 38                  15461                  15467                  15455
#> 39                  15828                  15834                  15827
#> 40                  16200                  16205                  16199
#> 41                  16559                  16564                  16558
#> 42                  16932                  16937                  16930
#> 43                  17295                  17301                  17289
#> 44                  17658                  17664                  17656
#>    transition_25_upper_ci transition_50_upper_ci threshold_10 threshold_25
#> 1                   14009                  14015      0.37851      0.39155
#> 2                   14369                  14375      0.37693      0.39080
#> 3                   14726                  14731      0.37742      0.38842
#> 4                   15103                  15108      0.37998      0.39287
#> 5                   15463                  15468      0.38284      0.39777
#> 6                   15831                  15836      0.38364      0.39496
#> 7                   16203                  16207      0.38298      0.39685
#> 8                   16563                  16567      0.38069      0.39491
#> 9                   16934                  16940      0.38404      0.40112
#> 10                  17298                  17303      0.38645      0.40147
#> 11                  17660                  17666      0.38568      0.39812
#> 12                  14009                  14015      0.37630      0.38882
#> 13                  14369                  14375      0.37443      0.38826
#> 14                  14727                  14732      0.37624      0.38887
#> 15                  15103                  15108      0.37793      0.39044
#> 16                  15463                  15469      0.38062      0.39279
#> 17                  15831                  15836      0.37979      0.39146
#> 18                  16203                  16207      0.38068      0.39289
#> 19                  16562                  16567      0.37902      0.39289
#> 20                  16934                  16940      0.38314      0.39781
#> 21                  17298                  17303      0.38318      0.39763
#> 22                  17661                  17666      0.38275      0.39709
#> 23                  14009                  14015      0.37281      0.38466
#> 24                  14369                  14375      0.37179      0.38485
#> 25                  14727                  14732      0.37360      0.38458
#> 26                  15104                  15110      0.37496      0.38846
#> 27                  15464                  15470      0.37757      0.38988
#> 28                  15831                  15837      0.37644      0.39131
#> 29                  16203                  16208      0.37749      0.39311
#> 30                  16562                  16567      0.37491      0.38630
#> 31                  16934                  16940      0.37985      0.39388
#> 32                  17298                  17303      0.37973      0.39314
#> 33                  17661                  17667      0.37985      0.39244
#> 34                  14009                  14014      0.37154      0.38420
#> 35                  14368                  14375      0.37030      0.38209
#> 36                  14727                  14732      0.37274      0.38519
#> 37                  15103                  15110      0.37414      0.38620
#> 38                  15464                  15470      0.37665      0.39029
#> 39                  15831                  15836      0.37569      0.38767
#> 40                  16203                  16208      0.37638      0.38867
#> 41                  16562                  16567      0.37401      0.38550
#> 42                  16934                  16940      0.37793      0.39412
#> 43                  17298                  17303      0.37917      0.39394
#> 44                  17661                  17666      0.37936      0.39340
#>    threshold_50 min_gcc max_gcc
#> 1       0.41656 0.36959 0.46330
#> 2       0.41848 0.36830 0.46187
#> 3       0.41786 0.36910 0.46467
#> 4       0.42230 0.37072 0.46848
#> 5       0.42735 0.37279 0.47880
#> 6       0.42595 0.37408 0.47168
#> 7       0.42850 0.37322 0.48011
#> 8       0.42144 0.37292 0.46627
#> 9       0.43168 0.37544 0.47953
#> 10      0.43702 0.37471 0.49464
#> 11      0.43244 0.37515 0.48051
#> 12      0.41571 0.36846 0.45856
#> 13      0.41436 0.36643 0.45576
#> 14      0.41841 0.36772 0.46124
#> 15      0.41641 0.36900 0.46265
#> 16      0.42339 0.37090 0.46819
#> 17      0.41919 0.37103 0.46533
#> 18      0.42278 0.37165 0.47384
#> 19      0.41839 0.37096 0.45962
#> 20      0.42780 0.37396 0.47496
#> 21      0.43070 0.37257 0.48293
#> 22      0.42475 0.37361 0.47361
#> 23      0.41136 0.36654 0.45121
#> 24      0.40917 0.36322 0.45227
#> 25      0.41164 0.36459 0.45679
#> 26      0.41545 0.36590 0.45970
#> 27      0.41930 0.36806 0.46548
#> 28      0.41741 0.36792 0.46155
#> 29      0.42319 0.36877 0.46818
#> 30      0.41468 0.36725 0.45493
#> 31      0.42277 0.37007 0.47035
#> 32      0.42395 0.36961 0.47438
#> 33      0.42560 0.37056 0.47053
#> 34      0.40848 0.36522 0.45095
#> 35      0.40978 0.36180 0.45102
#> 36      0.41278 0.36364 0.45603
#> 37      0.41573 0.36508 0.45839
#> 38      0.42077 0.36715 0.46527
#> 39      0.41419 0.36704 0.46021
#> 40      0.42430 0.36790 0.46794
#> 41      0.41429 0.36658 0.45438
#> 42      0.42361 0.36896 0.46988
#> 43      0.42552 0.36837 0.47689
#> 44      0.42060 0.36956 0.46906
#> 
#> $falling
#>    gcc_value transition_10 transition_25 transition_50 transition_10_lower_ci
#> 1     gcc_90         17842         17838         17828                  17843
#> 2     gcc_90         17474         17471         17462                  17475
#> 3     gcc_90         17102         17097         17090                  17103
#> 4     gcc_90         16736         16733         16725                  16737
#> 5     gcc_90         16364         16361         16354                  16365
#> 6     gcc_90         15996         15993         15984                  15998
#> 7     gcc_90         15632         15629         15621                  15633
#> 8     gcc_90         15275         15270         15257                  15276
#> 9     gcc_90         14900         14896         14889                  14901
#> 10    gcc_90         14540         14537         14530                  14541
#> 11    gcc_90         14176         14172         14166                  14177
#> 12    gcc_75         17841         17837         17827                  17843
#> 13    gcc_75         17473         17470         17461                  17474
#> 14    gcc_75         17100         17096         17088                  17101
#> 15    gcc_75         16735         16733         16726                  16736
#> 16    gcc_75         16364         16361         16354                  16365
#> 17    gcc_75         15996         15993         15985                  15997
#> 18    gcc_75         15630         15627         15617                  15631
#> 19    gcc_75         15275         15269         15257                  15276
#> 20    gcc_75         14900         14896         14888                  14901
#> 21    gcc_75         14540         14537         14530                  14541
#> 22    gcc_75         14175         14172         14164                  14176
#> 23    gcc_50         17838         17835         17827                  17839
#> 24    gcc_50         17473         17470         17461                  17474
#> 25    gcc_50         17100         17095         17087                  17101
#> 26    gcc_50         16733         16731         16724                  16734
#> 27    gcc_50         16363         16360         16353                  16364
#> 28    gcc_50         15996         15993         15985                  15997
#> 29    gcc_50         15628         15624         15615                  15629
#> 30    gcc_50         15274         15269         15257                  15276
#> 31    gcc_50         14900         14896         14887                  14901
#> 32    gcc_50         14540         14537         14530                  14541
#> 33    gcc_50         14175         14171         14165                  14176
#> 34  gcc_mean         17840         17836         17827                  17842
#> 35  gcc_mean         17473         17470         17461                  17474
#> 36  gcc_mean         17100         17096         17088                  17101
#> 37  gcc_mean         16735         16732         16725                  16736
#> 38  gcc_mean         16363         16360         16353                  16364
#> 39  gcc_mean         15996         15993         15985                  15997
#> 40  gcc_mean         15628         15624         15614                  15629
#> 41  gcc_mean         15275         15270         15258                  15276
#> 42  gcc_mean         14900         14896         14888                  14901
#> 43  gcc_mean         14540         14537         14529                  14541
#> 44  gcc_mean         14174         14171         14165                  14175
#>    transition_25_lower_ci transition_50_lower_ci transition_10_upper_ci
#> 1                   17840                  17832                  17839
#> 2                   17472                  17465                  17472
#> 3                   17099                  17092                  17100
#> 4                   16734                  16728                  16734
#> 5                   16362                  16356                  16362
#> 6                   15994                  15987                  15994
#> 7                   15630                  15623                  15630
#> 8                   15272                  15261                  15272
#> 9                   14897                  14892                  14897
#> 10                  14538                  14533                  14538
#> 11                  14173                  14168                  14174
#> 12                  17839                  17830                  17839
#> 13                  17471                  17463                  17471
#> 14                  17097                  17090                  17098
#> 15                  16734                  16728                  16734
#> 16                  16362                  16356                  16363
#> 17                  15994                  15987                  15994
#> 18                  15628                  15620                  15628
#> 19                  15271                  15260                  15272
#> 20                  14897                  14890                  14897
#> 21                  14538                  14532                  14538
#> 22                  14173                  14167                  14173
#> 23                  17836                  17829                  17836
#> 24                  17471                  17463                  17471
#> 25                  17096                  17089                  17098
#> 26                  16732                  16726                  16732
#> 27                  16361                  16355                  16362
#> 28                  15994                  15987                  15995
#> 29                  15625                  15618                  15626
#> 30                  15271                  15261                  15272
#> 31                  14897                  14890                  14898
#> 32                  14538                  14532                  14538
#> 33                  14172                  14167                  14173
#> 34                  17837                  17829                  17837
#> 35                  17471                  17463                  17471
#> 36                  17097                  17090                  17098
#> 37                  16733                  16727                  16733
#> 38                  16361                  16355                  16362
#> 39                  15994                  15987                  15995
#> 40                  15626                  15617                  15626
#> 41                  15271                  15262                  15272
#> 42                  14897                  14890                  14897
#> 43                  14538                  14532                  14538
#> 44                  14172                  14167                  14172
#>    transition_25_upper_ci transition_50_upper_ci threshold_10 threshold_25
#> 1                   17835                  17825      0.38425      0.39706
#> 2                   17469                  17459      0.38573      0.39839
#> 3                   17095                  17089      0.38827      0.40259
#> 4                   16731                  16723      0.38707      0.40014
#> 5                   16359                  16353      0.38289      0.39581
#> 6                   15991                  15981      0.38454      0.39649
#> 7                   15627                  15619      0.38056      0.39092
#> 8                   15267                  15254      0.38720      0.39961
#> 9                   14894                  14888      0.37714      0.38936
#> 10                  14536                  14529      0.37088      0.38549
#> 11                  14171                  14164      0.36925      0.38569
#> 12                  17835                  17825      0.38371      0.39554
#> 13                  17468                  17459      0.38561      0.39774
#> 14                  17094                  17087      0.39111      0.40175
#> 15                  16731                  16725      0.37750      0.38964
#> 16                  16359                  16353      0.38051      0.39294
#> 17                  15991                  15983      0.38049      0.39207
#> 18                  15624                  15615      0.38437      0.39466
#> 19                  15267                  15254      0.38391      0.39681
#> 20                  14894                  14887      0.37473      0.38627
#> 21                  14536                  14529      0.36932      0.38325
#> 22                  14170                  14163      0.37054      0.38180
#> 23                  17834                  17825      0.37673      0.39134
#> 24                  17468                  17459      0.38224      0.39368
#> 25                  17093                  17086      0.38579      0.39877
#> 26                  16729                  16723      0.38006      0.39056
#> 27                  16358                  16352      0.37972      0.39161
#> 28                  15991                  15983      0.37798      0.38902
#> 29                  15622                  15612      0.38366      0.39582
#> 30                  15267                  15255      0.38196      0.39184
#> 31                  14894                  14886      0.37150      0.38290
#> 32                  14536                  14528      0.36583      0.37889
#> 33                  14170                  14164      0.36625      0.38129
#> 34                  17834                  17825      0.37799      0.39058
#> 35                  17468                  17460      0.38202      0.39324
#> 36                  17094                  17087      0.38534      0.39640
#> 37                  16731                  16724      0.36938      0.38770
#> 38                  16358                  16352      0.37926      0.39114
#> 39                  15991                  15983      0.37780      0.38878
#> 40                  15622                  15611      0.38444      0.39657
#> 41                  15267                  15254      0.37964      0.39021
#> 42                  14894                  14887      0.37105      0.38236
#> 43                  14536                  14528      0.36496      0.37788
#> 44                  14170                  14164      0.36910      0.38082
#>    threshold_50 min_gcc max_gcc
#> 1       0.42160 0.37731 0.46360
#> 2       0.42288 0.37736 0.46806
#> 3       0.42484 0.38171 0.46580
#> 4       0.42332 0.38134 0.46286
#> 5       0.42570 0.37395 0.47024
#> 6       0.42086 0.37635 0.46386
#> 7       0.41729 0.37260 0.45963
#> 8       0.42218 0.38098 0.46126
#> 9       0.41421 0.37170 0.45127
#> 10      0.41243 0.36235 0.45758
#> 11      0.40800 0.36461 0.45029
#> 12      0.41868 0.37552 0.45902
#> 13      0.42116 0.37994 0.46118
#> 14      0.42395 0.38415 0.46014
#> 15      0.41624 0.37048 0.45724
#> 16      0.41994 0.37180 0.46331
#> 17      0.41532 0.37370 0.45648
#> 18      0.41755 0.37902 0.45461
#> 19      0.41653 0.37822 0.45353
#> 20      0.40967 0.36931 0.44670
#> 21      0.40806 0.36098 0.45208
#> 22      0.40762 0.36613 0.44412
#> 23      0.41406 0.37161 0.45564
#> 24      0.41795 0.37694 0.45761
#> 25      0.42047 0.38042 0.45636
#> 26      0.41580 0.37389 0.45386
#> 27      0.41576 0.37169 0.45819
#> 28      0.41124 0.37031 0.45086
#> 29      0.41435 0.37924 0.44930
#> 30      0.41201 0.37487 0.44715
#> 31      0.40619 0.36621 0.44255
#> 32      0.40263 0.35788 0.44718
#> 33      0.40112 0.36201 0.43948
#> 34      0.41340 0.37088 0.45435
#> 35      0.41690 0.37515 0.45621
#> 36      0.41832 0.37977 0.45485
#> 37      0.41267 0.36676 0.45294
#> 38      0.41603 0.37126 0.45746
#> 39      0.41067 0.37014 0.45096
#> 40      0.41465 0.37990 0.44804
#> 41      0.41030 0.37418 0.44586
#> 42      0.40440 0.36574 0.44158
#> 43      0.40343 0.35723 0.44566
#> 44      0.40145 0.36318 0.43684
#>