Calculates phenophases

This routine combines a forward and backward run of transition_dates function to calculate the phenophases in both rising and falling parts of a PhenoCam time series.

phenophases(data, mat, internal = TRUE, out_dir = tempdir(), ...)

Arguments

data
a PhenoCam data file (or data frame)

mat
mean annual temperature

internal
return PhenoCam data file or data frame

out_dir
output directory

...
pass parameters to the transition_dates() function

Value

Estimates of transition dates for both rising and falling parts of a PhenoCam time series. All time series are evaluated (gcc_90, gcc_75, etc). The function returns a nested list with UNIX time based values including uncertainties on these estimates and their associated thresholds. When written to disk UNIX dates are converted to YYYY-MM-DD. The nested list has named locations rising and falling, or location 1 and 2 in the list respectivelly.

Examples

# downloads a time series
download_phenocam(site = "harvard$",
                  veg_type = "DB",
                  roi_id = "1000",
                  frequency = "3")

#> Downloading: harvard_DB_1000_3day.csv

#> -- Flagging outliers!

#> -- Smoothing time series!

# read in data as data frame and calculate phenophases
df <- read_phenocam(file.path(tempdir(),"harvard_DB_1000_3day.csv"))
my_dates <- phenophases(df, internal = TRUE)

# print results
print(my_dates)

#> $rising
#>    gcc_value transition_10 transition_25 transition_50 transition_10_lower_ci
#> 1     gcc_90         14002         14007         14014                  14000
#> 2     gcc_90         14361         14366         14374                  14357
#> 3     gcc_90         14719         14724         14730                  14715
#> 4     gcc_90         15096         15101         15107                  15093
#> 5     gcc_90         15452         15460         15467                  15444
#> 6     gcc_90         15825         15828         15835                  15823
#> 7     gcc_90         16197         16200         16206                  16195
#> 8     gcc_90         16556         16560         16566                  16554
#> 9     gcc_90         16927         16932         16939                  16924
#> 10    gcc_90         17289         17295         17302                  17283
#> 11    gcc_90         17653         17658         17664                  17651
#> 12    gcc_90         18019         18027         18034                  18015
#> 13    gcc_75         14002         14007         14014                  14001
#> 14    gcc_75         14360         14366         14374                  14357
#> 15    gcc_75         14720         14725         14731                  14716
#> 16    gcc_75         15095         15101         15108                  15091
#> 17    gcc_75         15451         15461         15468                  15443
#> 18    gcc_75         15824         15829         15835                  15823
#> 19    gcc_75         16197         16201         16206                  16195
#> 20    gcc_75         16555         16560         16566                  16554
#> 21    gcc_75         16927         16933         16939                  16925
#> 22    gcc_75         17289         17295         17302                  17284
#> 23    gcc_75         17654         17659         17665                  17652
#> 24    gcc_75         18021         18028         18034                  18018
#> 25    gcc_50         14003         14007         14014                  14001
#> 26    gcc_50         14359         14366         14374                  14353
#> 27    gcc_50         14720         14725         14731                  14702
#> 28    gcc_50         15095         15101         15109                  15092
#> 29    gcc_50         15448         15462         15469                  15442
#> 30    gcc_50         15824         15829         15836                  15821
#> 31    gcc_50         16197         16201         16207                  16195
#> 32    gcc_50         16555         16560         16566                  16553
#> 33    gcc_50         16927         16933         16939                  16925
#> 34    gcc_50         17288         17295         17302                  17282
#> 35    gcc_50         17654         17659         17665                  17652
#> 36    gcc_50         18021         18028         18034                  18018
#> 37  gcc_mean         14002         14007         14014                  14001
#> 38  gcc_mean         14358         14366         14374                  14352
#> 39  gcc_mean         14719         14725         14731                  14703
#> 40  gcc_mean         15095         15101         15108                  15092
#> 41  gcc_mean         15448         15461         15469                  15442
#> 42  gcc_mean         15824         15829         15835                  15821
#> 43  gcc_mean         16197         16201         16206                  16195
#> 44  gcc_mean         16555         16560         16566                  16553
#> 45  gcc_mean         16926         16932         16939                  16924
#> 46  gcc_mean         17287         17295         17302                  17282
#> 47  gcc_mean         17653         17658         17665                  17651
#> 48  gcc_mean         18020         18027         18034                  18017
#>    transition_25_lower_ci transition_50_lower_ci transition_10_upper_ci
#> 1                   14006                  14012                  14004
#> 2                   14365                  14372                  14364
#> 3                   14723                  14728                  14722
#> 4                   15100                  15105                  15098
#> 5                   15459                  15465                  15456
#> 6                   15827                  15833                  15827
#> 7                   16199                  16204                  16199
#> 8                   16559                  16564                  16558
#> 9                   16931                  16937                  16930
#> 10                  17294                  17301                  17292
#> 11                  17657                  17662                  17656
#> 12                  18026                  18032                  18023
#> 13                  14006                  14012                  14004
#> 14                  14365                  14372                  14363
#> 15                  14724                  14729                  14723
#> 16                  15100                  15106                  15098
#> 17                  15460                  15466                  15456
#> 18                  15828                  15834                  15827
#> 19                  16200                  16205                  16199
#> 20                  16559                  16564                  16558
#> 21                  16932                  16937                  16930
#> 22                  17294                  17301                  17292
#> 23                  17658                  17663                  17656
#> 24                  18027                  18032                  18024
#> 25                  14006                  14012                  14005
#> 26                  14365                  14372                  14363
#> 27                  14724                  14730                  14723
#> 28                  15100                  15107                  15098
#> 29                  15461                  15467                  15456
#> 30                  15828                  15834                  15826
#> 31                  16200                  16205                  16199
#> 32                  16559                  16564                  16557
#> 33                  16932                  16938                  16930
#> 34                  17294                  17301                  17291
#> 35                  17658                  17664                  17657
#> 36                  18027                  18033                  18024
#> 37                  14006                  14012                  14005
#> 38                  14365                  14372                  14362
#> 39                  14724                  14729                  14722
#> 40                  15100                  15107                  15098
#> 41                  15460                  15467                  15455
#> 42                  15828                  15834                  15826
#> 43                  16200                  16205                  16199
#> 44                  16559                  16564                  16557
#> 45                  16931                  16937                  16929
#> 46                  17294                  17301                  17291
#> 47                  17657                  17663                  17656
#> 48                  18026                  18032                  18023
#>    transition_25_upper_ci transition_50_upper_ci threshold_10 threshold_25
#> 1                   14009                  14015      0.37778      0.39230
#> 2                   14369                  14375      0.37727      0.38876
#> 3                   14726                  14731      0.37749      0.39024
#> 4                   15103                  15108      0.37990      0.39440
#> 5                   15462                  15468      0.38294      0.39674
#> 6                   15831                  15836      0.38256      0.39343
#> 7                   16202                  16207      0.38363      0.39458
#> 8                   16562                  16567      0.38214      0.39265
#> 9                   16934                  16940      0.38501      0.39746
#> 10                  17297                  17303      0.38640      0.40218
#> 11                  17660                  17665      0.38393      0.39827
#> 12                  18029                  18035      0.38633      0.40085
#> 13                  14009                  14015      0.37572      0.38945
#> 14                  14368                  14375      0.37441      0.38635
#> 15                  14727                  14732      0.37623      0.38974
#> 16                  15102                  15109      0.37736      0.39175
#> 17                  15463                  15469      0.38015      0.39464
#> 18                  15831                  15836      0.37844      0.39324
#> 19                  16202                  16207      0.38099      0.39598
#> 20                  16562                  16567      0.37850      0.39068
#> 21                  16934                  16940      0.38236      0.39827
#> 22                  17297                  17303      0.38290      0.39786
#> 23                  17660                  17666      0.38270      0.39786
#> 24                  18029                  18035      0.38492      0.39911
#> 25                  14009                  14015      0.37480      0.38538
#> 26                  14369                  14375      0.37186      0.38322
#> 27                  14727                  14732      0.37360      0.38532
#> 28                  15103                  15110      0.37453      0.38667
#> 29                  15464                  15470      0.37721      0.39129
#> 30                  15831                  15837      0.37643      0.38894
#> 31                  16202                  16208      0.37775      0.39162
#> 32                  16562                  16567      0.37607      0.38769
#> 33                  16934                  16940      0.37928      0.39437
#> 34                  17297                  17303      0.37943      0.39385
#> 35                  17660                  17666      0.37938      0.39352
#> 36                  18029                  18035      0.38147      0.39506
#> 37                  14009                  14015      0.37129      0.38491
#> 38                  14368                  14375      0.37050      0.38306
#> 39                  14727                  14732      0.37259      0.38595
#> 40                  15103                  15109      0.37378      0.38701
#> 41                  15463                  15470      0.37675      0.38938
#> 42                  15831                  15836      0.37574      0.38941
#> 43                  16202                  16207      0.37686      0.39141
#> 44                  16562                  16567      0.37510      0.38705
#> 45                  16934                  16940      0.37774      0.39068
#> 46                  17297                  17303      0.37848      0.39453
#> 47                  17660                  17666      0.37782      0.39017
#> 48                  18029                  18035      0.38025      0.39261
#>    threshold_50 min_gcc max_gcc
#> 1       0.41959 0.36939 0.46404
#> 2       0.41847 0.36807 0.46226
#> 3       0.41821 0.36904 0.46474
#> 4       0.42145 0.37072 0.46791
#> 5       0.42737 0.37257 0.47722
#> 6       0.42526 0.37278 0.47088
#> 7       0.42918 0.37310 0.47993
#> 8       0.42025 0.37282 0.46639
#> 9       0.43000 0.37542 0.47868
#> 10      0.43546 0.37452 0.49377
#> 11      0.42962 0.37511 0.47863
#> 12      0.43156 0.37698 0.47825
#> 13      0.41534 0.36824 0.45799
#> 14      0.41440 0.36632 0.45654
#> 15      0.41732 0.36770 0.46140
#> 16      0.42007 0.36887 0.46274
#> 17      0.42234 0.37066 0.46851
#> 18      0.41914 0.37094 0.46506
#> 19      0.42341 0.37160 0.47269
#> 20      0.41726 0.37114 0.45955
#> 21      0.42614 0.37392 0.47379
#> 22      0.42935 0.37253 0.48223
#> 23      0.42595 0.37358 0.47124
#> 24      0.42608 0.37578 0.47128
#> 25      0.41109 0.36625 0.45179
#> 26      0.40938 0.36332 0.45276
#> 27      0.41078 0.36449 0.45668
#> 28      0.41516 0.36575 0.45903
#> 29      0.41893 0.36802 0.46492
#> 30      0.41703 0.36776 0.46090
#> 31      0.42177 0.36873 0.46793
#> 32      0.41363 0.36744 0.45467
#> 33      0.42140 0.37009 0.46980
#> 34      0.42290 0.36946 0.47371
#> 35      0.42051 0.37074 0.46771
#> 36      0.42047 0.37227 0.46584
#> 37      0.41148 0.36482 0.45093
#> 38      0.40999 0.36200 0.45164
#> 39      0.41188 0.36350 0.45557
#> 40      0.41146 0.36501 0.45748
#> 41      0.42018 0.36708 0.46473
#> 42      0.41400 0.36689 0.45974
#> 43      0.41812 0.36782 0.46760
#> 44      0.41319 0.36679 0.45457
#> 45      0.42199 0.36879 0.46919
#> 46      0.42440 0.36839 0.47613
#> 47      0.42180 0.36962 0.46640
#> 48      0.42146 0.37130 0.46658
#> 
#> $falling
#>    gcc_value transition_10 transition_25 transition_50 transition_10_lower_ci
#> 1     gcc_90         17840         17837         17826                  17842
#> 2     gcc_90         17473         17469         17460                  17474
#> 3     gcc_90         17102         17098         17090                  17103
#> 4     gcc_90         16736         16733         16726                  16737
#> 5     gcc_90         16365         16361         16355                  16366
#> 6     gcc_90         15996         15992         15982                  15997
#> 7     gcc_90         15632         15628         15621                  15634
#> 8     gcc_90         15275         15270         15258                  15276
#> 9     gcc_90         14900         14896         14889                  14901
#> 10    gcc_90         14540         14537         14530                  14541
#> 11    gcc_90         14176         14173         14166                  14177
#> 12    gcc_75         17840         17836         17825                  17841
#> 13    gcc_75         17473         17469         17461                  17474
#> 14    gcc_75         17102         17098         17089                  17103
#> 15    gcc_75         16735         16733         16727                  16736
#> 16    gcc_75         16365         16361         16354                  16366
#> 17    gcc_75         15996         15992         15984                  15997
#> 18    gcc_75         15630         15627         15619                  15631
#> 19    gcc_75         15275         15269         15257                  15276
#> 20    gcc_75         14899         14895         14887                  14900
#> 21    gcc_75         14540         14537         14530                  14541
#> 22    gcc_75         14175         14172         14165                  14176
#> 23    gcc_50         17839         17835         17828                  17840
#> 24    gcc_50         17473         17469         17461                  17474
#> 25    gcc_50         17100         17096         17088                  17101
#> 26    gcc_50         16733         16731         16725                  16734
#> 27    gcc_50         16363         16360         16353                  16364
#> 28    gcc_50         15996         15992         15983                  15997
#> 29    gcc_50         15629         15625         15616                  15630
#> 30    gcc_50         15274         15269         15258                  15276
#> 31    gcc_50         14900         14896         14887                  14901
#> 32    gcc_50         14540         14536         14529                  14541
#> 33    gcc_50         14175         14172         14165                  14176
#> 34  gcc_mean         17839         17836         17826                  17841
#> 35  gcc_mean         17473         17469         17461                  17474
#> 36  gcc_mean         17100         17097         17089                  17101
#> 37  gcc_mean         16735         16732         16726                  16736
#> 38  gcc_mean         16364         16360         16353                  16365
#> 39  gcc_mean         15996         15992         15983                  15997
#> 40  gcc_mean         15629         15626         15617                  15630
#> 41  gcc_mean         15275         15270         15258                  15276
#> 42  gcc_mean         14900         14895         14887                  14901
#> 43  gcc_mean         14540         14536         14529                  14541
#> 44  gcc_mean         14175         14172         14166                  14176
#>    transition_25_lower_ci transition_50_lower_ci transition_10_upper_ci
#> 1                   17838                  17829                  17838
#> 2                   17470                  17463                  17471
#> 3                   17099                  17092                  17100
#> 4                   16734                  16728                  16734
#> 5                   16362                  16356                  16363
#> 6                   15993                  15986                  15993
#> 7                   15629                  15623                  15630
#> 8                   15272                  15261                  15272
#> 9                   14897                  14891                  14897
#> 10                  14538                  14532                  14538
#> 11                  14174                  14169                  14174
#> 12                  17837                  17828                  17837
#> 13                  17470                  17463                  17471
#> 14                  17099                  17091                  17100
#> 15                  16734                  16729                  16734
#> 16                  16362                  16356                  16363
#> 17                  15993                  15987                  15994
#> 18                  15628                  15621                  15628
#> 19                  15271                  15261                  15272
#> 20                  14896                  14890                  14897
#> 21                  14538                  14532                  14538
#> 22                  14173                  14167                  14173
#> 23                  17837                  17830                  17837
#> 24                  17470                  17463                  17471
#> 25                  17097                  17090                  17098
#> 26                  16732                  16727                  16732
#> 27                  16361                  16355                  16362
#> 28                  15993                  15986                  15994
#> 29                  15626                  15619                  15627
#> 30                  15271                  15261                  15272
#> 31                  14897                  14889                  14898
#> 32                  14537                  14532                  14538
#> 33                  14173                  14167                  14173
#> 34                  17837                  17829                  17837
#> 35                  17470                  17463                  17471
#> 36                  17098                  17091                  17098
#> 37                  16733                  16728                  16733
#> 38                  16361                  16355                  16362
#> 39                  15993                  15986                  15994
#> 40                  15627                  15620                  15628
#> 41                  15271                  15261                  15272
#> 42                  14897                  14889                  14897
#> 43                  14537                  14532                  14538
#> 44                  14173                  14168                  14173
#>    transition_25_upper_ci transition_50_upper_ci threshold_10 threshold_25
#> 1                   17834                  17823      0.39118      0.40185
#> 2                   17467                  17458      0.38944      0.40499
#> 3                   17096                  17089      0.38669      0.39957
#> 4                   16731                  16724      0.38761      0.39827
#> 5                   16359                  16354      0.38018      0.39616
#> 6                   15990                  15975      0.38633      0.39936
#> 7                   15626                  15620      0.38093      0.39416
#> 8                   15267                  15255      0.38694      0.39965
#> 9                   14894                  14888      0.37730      0.39007
#> 10                  14535                  14528      0.37131      0.38478
#> 11                  14171                  14165      0.36781      0.38048
#> 12                  17833                  17823      0.38691      0.39969
#> 13                  17467                  17459      0.38516      0.40032
#> 14                  17096                  17088      0.38426      0.39648
#> 15                  16731                  16726      0.37746      0.38710
#> 16                  16359                  16353      0.37800      0.39349
#> 17                  15990                  15981      0.38221      0.39494
#> 18                  15625                  15617      0.38296      0.39325
#> 19                  15267                  15255      0.38368      0.39675
#> 20                  14893                  14886      0.37740      0.38931
#> 21                  14535                  14528      0.36967      0.38246
#> 22                  14170                  14164      0.36927      0.38079
#> 23                  17833                  17826      0.37558      0.39079
#> 24                  17467                  17459      0.38176      0.39632
#> 25                  17094                  17087      0.38493      0.39712
#> 26                  16729                  16724      0.37838      0.38920
#> 27                  16358                  16352      0.38067      0.39133
#> 28                  15990                  15981      0.37962      0.39179
#> 29                  15623                  15614      0.37983      0.39161
#> 30                  15266                  15255      0.38178      0.39199
#> 31                  14893                  14885      0.37168      0.38313
#> 32                  14535                  14528      0.36607      0.38138
#> 33                  14170                  14164      0.36520      0.37659
#> 34                  17833                  17824      0.38213      0.39220
#> 35                  17467                  17459      0.38150      0.39582
#> 36                  17095                  17087      0.38451      0.39428
#> 37                  16731                  16725      0.36984      0.38513
#> 38                  16358                  16352      0.37651      0.39083
#> 39                  15990                  15980      0.37936      0.39149
#> 40                  15624                  15615      0.38043      0.38968
#> 41                  15267                  15255      0.37944      0.39029
#> 42                  14893                  14885      0.37111      0.38528
#> 43                  14535                  14528      0.36511      0.38045
#> 44                  14170                  14165      0.36438      0.37597
#>    threshold_50 min_gcc max_gcc
#> 1       0.42484 0.38458 0.46464
#> 2       0.42726 0.38395 0.46874
#> 3       0.42561 0.38063 0.46479
#> 4       0.42157 0.38164 0.46135
#> 5       0.42222 0.37397 0.46870
#> 6       0.42225 0.38031 0.46338
#> 7       0.41758 0.37278 0.45937
#> 8       0.42128 0.38105 0.46107
#> 9       0.41376 0.37164 0.45374
#> 10      0.41232 0.36511 0.45733
#> 11      0.40812 0.36105 0.44984
#> 12      0.42221 0.38232 0.45970
#> 13      0.42153 0.37997 0.46147
#> 14      0.42137 0.37843 0.46015
#> 15      0.41392 0.36954 0.45652
#> 16      0.42067 0.37196 0.46364
#> 17      0.41643 0.37646 0.45580
#> 18      0.41520 0.37542 0.45368
#> 19      0.41690 0.37829 0.45273
#> 20      0.41149 0.37067 0.44902
#> 21      0.40813 0.36364 0.45109
#> 22      0.40514 0.36278 0.44441
#> 23      0.41312 0.36956 0.45596
#> 24      0.41804 0.37678 0.45767
#> 25      0.41792 0.37778 0.45618
#> 26      0.41262 0.37092 0.45340
#> 27      0.41635 0.37246 0.45794
#> 28      0.41382 0.37387 0.45129
#> 29      0.41304 0.37532 0.44898
#> 30      0.41100 0.37481 0.44709
#> 31      0.40554 0.36619 0.44420
#> 32      0.40483 0.36043 0.44553
#> 33      0.40146 0.35891 0.43957
#> 34      0.41546 0.37410 0.45467
#> 35      0.41704 0.37650 0.45645
#> 36      0.41586 0.37716 0.45425
#> 37      0.41023 0.36444 0.45210
#> 38      0.41672 0.37067 0.45702
#> 39      0.41308 0.37364 0.45129
#> 40      0.41091 0.37324 0.44768
#> 41      0.41044 0.37426 0.44565
#> 42      0.40601 0.36568 0.44452
#> 43      0.40338 0.35958 0.44428
#> 44      0.39860 0.35804 0.43720
#>